+++ src/hooks/index.ts
@@ -2,6 +2,7 @@ export { default as useEventNavigate } from './useEventNavigate';
 export { default as useFilterContext } from './useFilterContext';
 export { default as useFrameContext } from './useFrameContext';
 export { default as useIsMobileDevice } from './useIsMobileDevice';
+export { default as usePhotoDownload } from './usePhotoDownload';
 export { default as usePhotosContext } from './usePhotosContext';
 export { default as useTakePhoto } from './useTakePhoto';
 export { default as useToasterRiv } from './useToasterRiv';
